requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
blockbuster==1.5.24
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cohere==5.15.0
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
fastavro==1.11.1
filelock==3.18.0
forbiddenfruit==0.1.4
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.44
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
idna==3.10
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-rs==0.29.1
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langgraph==0.5.0
langgraph-api==0.2.67
langgraph-checkpoint==2.1.0
langgraph-cli==0.3.3
langgraph-prebuilt==0.5.1
langgraph-runtime-inmem==0.3.3
langgraph-sdk==0.1.72
langsmith==0.4.4
markupsafe==3.0.2
marshmallow==3.26.1
multidict==6.5.0
mypy-extensions==1.1.0
narwhals==1.44.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.3.1
openai==1.90.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
pdf2image==1.17.0
pillow==11.2.1
pinecone==7.2.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.0
pydeck==0.9.1
pyjwt==2.10.1
pymupdf==1.26.1
pypdf==5.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.1.3
starlette==0.47.1
streamlit==1.46.1
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
truststore==0.10.1
types-requests==2.32.4.20250611
typhoon-ocr==0.3.8
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
